extention functions added in stats package
