+++ app/src/main/java/com/iranmobiledev/moodino/ui/MainActivity.kt
@@ -80,7 +80,6 @@ class MainActivity : AppCompatActivity() {
                 putString("time", dateContainer.time())
             })
         }
-
     }
 
     private fun initViews() {
